+++ src/dockerblade/daemon.py
@@ -48,8 +48,9 @@ class DockerDaemon:
     def provision(self, image: str) -> Container:
         """Creates a Docker container from a given image."""
         logger.debug(f"provisioning container for image [{image}]")
-        container: Container
-        raise NotImplementedError
+        docker_container = \
+            self.client.containers.run(image, stdin_open=True, detach=True)
+        container = self.attach(docker_container.id)
         logger.debug(f"provisioned container [{container}]"
                      f" for image [{image}]")
         return container